requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.3
cachetools==2.1.0
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0
django-jet==1.0.7
django-letsencrypt==3.0.1
django-maintenance-mode==0.11.0
eth-typing==2.0.0
gevent==1.2.0
gevent-eventemitter==2.0
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
greenlet==0.4.15
httplib2==0.11.3
idna==2.7
oauth2client==4.1.2
oauthlib==2.1.0
protobuf==3.6.1
psycopg2==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
six==1.11.0
social-auth-app-django==2.1.0
git+https://github.com/Phalelashvili/social-core.git#egg=social-auth-core
steam==0.9.0
git+https://github.com/Phalelashvili/steampy.git#egg=steampy
toolz==0.9.0
uritemplate==3.0.0
urllib3==1.23
vdf==2.3
django-webpush
git+https://github.com/Phalelashvili/emoneyge-py
termcolor
pyotp
redis
psutil